requirements.txt
Flask
ultralytics
Pillow
pandas